[docs] Getting Started with TVM: Auto Scheduler and matmul (#7644)
Moves the auto scheduler with matmul example into the tutorial,
expands to follow the flow of the larger getting started tutorial.
Indended to follow the AutoTVM tutorial on matrix multiplication.
