+++ docs/conf.py
@@ -210,10 +210,11 @@ subsection_order = ExplicitOrder(
 # The unlisted files always appear after listed files.
 within_subsection_order = {
     "get_started": [
-        "relay_quick_start.py",
-        "tensor_expr_get_started.py",
         "tvmc_command_line_driver.py",
+        "tensor_expr_get_started.py",
+        "autoschedule_matmul.py",
         "cross_compilation_and_rpc.py",
+        "relay_quick_start.py",
     ],
     "frontend": [
         "from_pytorch.py",
